Add wiki links plugin for all types of objects
Regardless of use of Wiki plugin, one can use this plugin to add links to objects inside the same project.
For example, the project overview can include things like [message:3] that will be changed to a link to message of ID 3 (only if part of same project)

Works with:
[message:6]
[task_list:1]
[ticket:1]
[milestone:1]
[file:1]

Add plugin hooks to trigger the filters.